CUR, loading and ensembles (#5)
* CUR improvements

* CUR truncation in configuration set

* Added trajectory from ASE minimisation

* Fixed type in configuration _list

* GAP optimisations

* Added GAP MD

* Minor fixes

* Added thresholding on energies/forces

* More flexible plotting

* Simple error estimation with a committee model

* Started test on ensemble error estimate

* Added data load without system

* Option to use all training data in ensemble

* Added remove highest energy configurations

+++ gaptrain/md.py
@@ -125,7 +125,8 @@ def run_gapmd(configuration, gap, temp, dt, interval, **kwargs):
     os.environ['OMP_NUM_THREADS'] = str(n_cores)
     logger.info(f'Using {n_cores} cores for GAP MD')
 
-    assert os.path.exists(f'{gap.name}.xml')
+    if not os.path.exists(f'{gap.name}.xml'):
+        raise IOError('GAP parameter file (.xml) did not exist')
 
     # Print a Python script to execute quippy and use ASE to drive the dynamics
     with open(f'gap.py', 'w') as quippy_script:
@@ -136,15 +137,15 @@ def run_gapmd(configuration, gap, temp, dt, interval, **kwargs):
               'from ase.md.velocitydistribution import MaxwellBoltzmannDistribution',
               'from ase import units',
               'from ase.md.langevin import Langevin',
-              f'system = read("config.xyz")',
+              'system = read("config.xyz")',
               f'system.cell = [{a}, {b}, {c}]',
               'system.pbc = True',
               'system.center()',
-              f'pot = quippy.Potential("IP GAP", \n'
+              'pot = quippy.Potential("IP GAP", \n'
               f'                      param_filename="{gap.name}.xml")',
               'system.set_calculator(pot)',
               f'MaxwellBoltzmannDistribution(system, {temp} * units.kB)',
-              f'traj = Trajectory("tmp.traj", \'w\', system)\n'
+              'traj = Trajectory("tmp.traj", \'w\', system)\n'
               f'dyn = Langevin(system, {dt:.1f} * units.fs, {temp} * units.kB, 0.02)',
               f'dyn.attach(traj.write, interval={interval})',
               f'dyn.run(steps={n_steps})',
